requirements.txt
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cysystemd==1.5.4
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
secure==0.3.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
-e git+https://github.com/costa-rica/ServerStatusWebsite_modules.git@960302e406128b50ef54aa6c644f02814eabe0f0#egg=ss_modules&subdirectory=ss_modules
systemd==0.17.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.6
